Fixed bugreport:7105 where pierce defense must ignore def and add (def/2) as additional ATK damage. Fixed bugreport:7148 where r17072 broke the ability of smiths to forge weapons.
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@17089 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/map/clif.h
@@ -442,7 +442,7 @@ void clif_skill_estimation(struct map_session_data *sd,struct block_list *dst);
 void clif_skill_warppoint(struct map_session_data* sd, uint16 skill_id, uint16 skill_lv, unsigned short map1, unsigned short map2, unsigned short map3, unsigned short map4);
 void clif_skill_memomessage(struct map_session_data* sd, int type);
 void clif_skill_teleportmessage(struct map_session_data *sd, int type);
-void clif_skill_produce_mix_list(struct map_session_data *sd, uint16 skill_id, int trigger);
+void clif_skill_produce_mix_list(struct map_session_data *sd, int skill_id, int trigger);
 void clif_cooking_list(struct map_session_data *sd, int trigger, uint16 skill_id, int qty, int list_type);
 
 void clif_produceeffect(struct map_session_data* sd,int flag,int nameid);
